Get params now returns a device flags word in [e]ax.
Int 33h now clears CF on entry and removed all cases of clearing before exiting.

deviceChanged now returns status like XT-BIOS (i.e. AH=06h and CF=CY if device since swapped).

busScan now returns AH=06h and CF=CY and thence exits the Int33h call if the port is empty (and will reset the line if it is set).
Similarly exits the call if ANY critical failures occur.

busScan will NOT return AH=06h and CF=CY if a device swap occured but the port is not empty but the swap will occur silently and there will be no record that the swap occured.